v0.1.2 Scene Manager Rework
The Scene Manager, rather than a Singleton, is now a Scriptable Object. This allows us to call its methods without needing to have an object in the scene that contains the script, and we only need a reference to this Scriptable Object if we want to access any of its methods. This means we can even call its methods from buttons!

UI Scaling in the Sample Scene was fixed.

The PlayerDataManager Singleton was fixed to actually be a singleton.

Unnecessary variables from the Main Menu Manager Script were removed.