v0.2.0 Save System
The game is now prepared to save data to a file and load from it. As of now, the data is not encrypted, so it can be easily modified.

PlayerDataManager was renamed to GameManager and SceneManagerScript was renamed to SceneManager.

The MainMenuManager Script was removed because its contents were moved to the GameManager Script.