v0.3.0 Currency Manager & Leaderboard
Note to self: Working with save files is a pain.

The Main Menu now has a functioning leaderboard, which loads data from a list containing the top 10 players ranked by points (sorting and adding players aren not implemented yet). When closing the program, it also saves the data to the save file.

The currency manager uses events to send a message when the amount of credits has changed, and the new Main Menu UI Manager updates the text accordingly, although there still isn't a way to get credits other than setting them manually thorugh the inspector or the save file.