v0.4.0 Fading Transition Between Scenes
When pressing the play button of the Main Menu, the menu becomes transparent and the background fades to black to allow for a smoother transition.

Similarly, as long as a Fade Manager Object exists in a scene, the scene will fade in automatically when loaded.

The Fade Manager has a method to load scenes after fading.

A prefab was created for this Fade Manager.