v0.4.1 Main Menu Leaderboard Fix
The Leaderboard of the Main Menu should now work as intended by loading the data of the leaderboard each time the script starts, rather than when loading the save data the first time.

Made the UpdateLeaderboard method private since we only need to call it from inside the same script.

Updated default leaderboard text in case loading data fails. This may be changed in the future.

Also updated the build settings.