v0.6.0 Ship Selector & Main Menu Ship
The Ship now gets instantiated in the Main Menu, but it doesn't in the playing scene since I need to add a script for it.

Added a menu to change the ship settings in-game. The ship changes its type and color in real time by instantiating and destroying the corresponding new ship and old ship.