v0.6.2a Ship & Color Unlock
The ship and colors can now be unlocked through code. Rather than dictionaries, we now use lists, whose data can be easily stored in files.

In case it's the first time executing the program, we initialize the lists and current ship to a default state to ensure no nullreferences.

Rather than enums, we now pass integers as values to lock and unlock types and colors.

The Main Menu method to determine which types or colors are locked is still in the update function, which should be optimized in the future. Additionally, this method has been divided so its methods can be called individually or at the same time.

The player's save data and leaderboard data have been separated into different files.