v0.6.3 Leaderboard Sorting
The leaderboard is now sorted before loading it, ensuring that the player with the most points stays at the top of the ranking.

Reduced the font size of the leaderboard so more characters can fit in a single line.

Removed unnecessary code when managing the color locks since they don't have any children objects.