v0.8.1 Movement Rotation
The player now rotates slightly in the direction of movement when moving. Similarly to movement lerping, this is handled with coroutines.

We also keep track of the previous point index to determine if we should start moving, or to not move if the player is already at the edge and tries to move further. It is also used as a way to calculate the direction of rotation.