v0.8.3 Camera Position Change
The camera can now change its position by changing the active Cinemachine Camera based on a simple Animator that determines if we are in Horizontal or Vertical Mode.

For testing purposes, an InvokeRepeating method calls this ChangePerspective method every 5 seconds. It can also be called by clicking on the corresponding checkbox in the inspector.