v0.8.3a Perspective Player Position Reset
The method to change perspective was changed from a normal method to a coroutine with a loop. It now forces the perspective delay to wait until the player isn't moving to change.

The player is now notified to center its position at 75% of the camera animation (0.75 seconds).

A boolean was added to the player movement lerping coroutine to ensure that variables aren't modified by this script while the camera is moving.