v0.11.0 Player Collision Detection
The Player can now detect collision with asteroids with a rigidbody and box collider. Each ship has its own different dimensions for the collider.

Asteroids return to the pool upon colliding.

When changing perspective, currently active asteroids move to the side to not hit the player.

Asteroid prefabs now also include a rigidbody.