v0.15.0 Motion Controls
On Android, the player now moves the spaceship by rotating their device.

The previous movement is still preserved in the Player Movement Script, as well as a backup of the Play Scene with previous button placements.

A debug scene was added to test the accelerometer and gyroscope. For now, it is accessed through the hidden Credits button.