v0.16.1 Item update 2
Items are now mostly functional, except for the shield. Its function still has to be implemented.

The chance for portals to launch items has been decreased to 1/5.

Several methods checking for time have been updated so that they use unscaled time in order for them to work as intended in slowed down time.