v0.16.2 Item Shield Update
The shield item now works, spawning a shield around the spaceship when collected. The shield lasts 10 seconds (scaled with Time.timeScale) and blocks everything as if it was a projectile. This also means the player can't collect other items while they have a shield on.

Added the option to set the prefab from the InGameManager script through the inspector.

The shield has several values that can be set through the inspector: the scale that it should have when spawned depending on the current ship type.

All of the item power ups have been divided into their own coroutines, allowing them to have different timers.

Items now last 4.5 seconds in game if the player doesn't collect them, which should make them less frequent.

While in slowed down time, the player now moves as normal. This was done by changing the Time.deltaTime to Time.unscaledDeltaTime, which is closer to the intended functionality than before.