v0.16.3 Motion Controls Settings
The player can now choose if they want to enable or disable Motion Controls through the settings, although this only works on mobile.

The In-Game UI Manager was updated so that it can now use a divided canvas for motion and no-motion. This means that a different canvas will be activated depending on the settings ONLY if the player is on mobile (android). Otherwise, the minimal canvas is used.

Also updated the save system so that the motion preferences are saved.