v0.16.4 Performance Optimization & Bugfixes
The camera now uses Dynamic Resolution to reach the desired frame rate in case it is not possible to reach them normally, up to a factor of 0.5 (settable through the inspector). As of now, there is no option to turn this off in-game.

Fixed a bug wherre the player would constantly get added to the leaderboard upon the game ending. This was fixed by changing activeSelf for activeInHierarchy inside InGameUIManager.

Fade animation now uses unscaled deltaTime for fading out.

We now instantiate 50 projectiles upon the game starting. This is not that many considering I was able to get 157 while testing.

Fixed the shield not completely covering the Bird by increasing its scale.

Asteroid collision detection was changed to Continuous Dynamic and projectile colliders were made a bit bigger.

The player should no longer be able to shoot while dead if they were shooting right before dying.