v0.16.5 Performance Settings & Billboarding
The settings menu has been adjusted. We now use dropdown menus for most things.

Dynamic resolution is now a setting that the player can choose by changing the Graphics Mode in settings. It is also saved to a file, as well as motion controls.

Billboarding was added for item icons to always face the camera.